requirements.txt
dj-database-url==0.5.0
django==2.1.5
django-heroku==0.3.1
gunicorn==19.9.0
psycopg2-binary==2.7.7
pytz==2018.9
whitenoise==4.1.2
djangorestframework==3.9.1
markdown==3.0.1
django-filter==2.1.0
